refactor: better cleanup and separation of components
