+++ src/defs.ts
@@ -24,6 +24,7 @@ export * from "./types/task";
 export * from "./types/taskMiddleware";
 export * from "./types/resourceMiddleware";
 export * from "./types/meta";
+export * from "./types/runner";
 
 // Useful other types that are kind-of spread out.
 export { ICacheInstance } from "./globals/middleware/cache.middleware";
