+++ src/index.ts
@@ -49,5 +49,4 @@ export * as definitions from "./defs";
 export * from "./models";
 export * from "./globals/types";
 export * as Errors from "./errors";
-export { RunOptions } from "./run";
 export { Context } from "./context";
