+++ src/run.ts
@@ -5,12 +5,10 @@ import { EventManager } from "./models/EventManager";
 import { globalEvents } from "./globals/globalEvents";
 import { Store } from "./models/Store";
 import { findCircularDependencies } from "./tools/findCircularDependencies";
-import { CircularDependenciesError, RuntimeError } from "./errors";
-import { globalResources } from "./globals/globalResources";
-import { Logger, LogLevels, PrintStrategy } from "./models/Logger";
-import { ResourceNotFoundError } from "./errors";
+import { CircularDependenciesError } from "./errors";
+import { Logger } from "./models/Logger";
 import { isResourceWithConfig } from "./define";
-import { debugResource, DebugFriendlyConfig } from "./globals/resources/debug";
+import { debugResource } from "./globals/resources/debug";
 import {
   registerProcessLevelSafetyNets,
   registerShutdownHook,
@@ -19,44 +17,9 @@ import {
   OnUnhandledError,
   createDefaultUnhandledError,
   bindProcessErrorHandler,
-  safeReportUnhandledError,
 } from "./models/UnhandledError";
 import { RunResult } from "./models/RunResult";
-
-export type RunOptions = {
-  /**
-   * Defaults to undefined. If true, we introduce logging to the console.
-   */
-  debug?: DebugFriendlyConfig;
-  logs?: {
-    /**
-     * Defaults to info. Use null to disable logging.
-     */
-    printThreshold?: null | LogLevels;
-    /**
-     * Defaults to PRETTY. How to print the logs.
-     */
-    printStrategy?: PrintStrategy;
-    /**
-     * Defaults to false. If true, we buffer logs until the root resource is ready.
-     * This provides you with the chance to see the logs before the root resource is ready.
-     */
-    bufferLogs?: boolean;
-  };
-  /**
-   * When true (default), installs a central error boundary that catches uncaught errors
-   * from process-level events and routes them to `onUnhandledError`.
-   */
-  errorBoundary?: boolean;
-  /**
-   * When true (default), installs SIGINT/SIGTERM handlers that call dispose() on the root.
-   */
-  shutdownHooks?: boolean;
-  /**
-   * Custom handler for any unhandled error caught by Runner. Defaults to logging via the created logger.
-   */
-  onUnhandledError?: OnUnhandledError;
-};
+import { RunOptions } from "./types/runner";
 
 /**
  * This is the central function that kicks off you runner. You can run as many resources as you want in a single process, they will run in complete isolation.
